Send the OpenCV data to the AgentScript model
We perform a very simple calculation to get a rough center-of-mass
of the squares and triangles, and then send that data to the
model to automatically place the windfarms and villages.

No smoothing of data or anything else is being done at this point,
so the AS objects jump around if the video feed is noisy.