Add simple implementation of power-lines
This is an inefficient way of detecting whether a group
of powerlines (which will be sent by openCV as a single
contour group) touches both a village and a windfarm. If
it does, then it assumes that power is being delivered to
village.

The shapes-client is updated as well, now sending only
red non-polygons, which are interpreted as powerlines.

Next, we need to group our individual calibrations and
find them all independently.