Model: Add very simple serialization demo
This simply JSON encodes the agents in the model
(excluding the powerlines, for now), and btoa-encodes
them.

If we load in a model with the encoded data in the hash,
we parse the data and add the agents, and then (for now)
block the OpenCV data as that will destroy the loaded agents.

In the future, we would want to save the data to a database,
instead of encoding it in the url, naturally.